изменения после ревью

+++ backend/foodgram/users/views.py
@@ -2,41 +2,82 @@ from rest_framework import status
 from rest_framework.generics import ListAPIView, get_object_or_404
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from rest_framework import viewsets
+from rest_framework.decorators import action
+
 from users.permissions import AuthorOrReadonly
 
 from .models import Follow, User
 from .serializers import FollowingSerializer, FollowSerializer
 
 
-class SubscriptionsAPIView(ListAPIView):
-    """API вью для работы с подпиской на авторов"""
+class SubscribeViewSet(viewsets.ModelViewSet):
+    """API вью для подписки и отписки от авторов"""
     serializer_class = FollowingSerializer
     permission_classes = (AuthorOrReadonly, )
 
     def get_queryset(self):
         return Follow.objects.filter(user=self.request.user)
 
+    @action(detail=False,
+            url_path='subscriptions',
+            methods=['get'],
+            permission_classes=[AuthorOrReadonly],
+            )
+    def subscriptions(self, request):
+        follow = Follow.objects.filter(user=self.request.user)
+        serializer = self.get_serializer(follow, many=True)
+        return Response(serializer.data)
 
-class SubscribeAPIView(APIView):
-    """API вью для подписки и отписки от авторов"""
-    permission_classes = (AuthorOrReadonly, )
-
-    def post(self, request, pk=None):
+    @action(detail=True,
+            url_path='subscribe',
+            methods=['post', 'delete'],
+            permission_classes=[AuthorOrReadonly],
+            )
+    def subscribe(self, request, pk=None):
         author = get_object_or_404(User, pk=pk)
-        data = {'user': request.user.id, 'author': author.id}
-        serializer = FollowSerializer(
-            data=data, context={'request': request}
-        )
-        if serializer.is_valid():
+        if request.method == 'POST':
+            data = {'user': request.user.id, 'author': author.id}
+            serializer = FollowSerializer(
+                data=data, context={'request': request}
+            )
+            serializer.is_valid(raise_exception=True)
             serializer.save()
-            return Response(serializer.data,
-                            status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request, pk):
-        author = get_object_or_404(User, pk=pk)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
         follow = get_object_or_404(
             Follow, user=request.user, author=author
         )
         follow.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+# class SubscriptionsAPIView(ListAPIView):
+#     """API вью для работы с подпиской на авторов"""
+#     serializer_class = FollowingSerializer
+#     permission_classes = (AuthorOrReadonly, )
+
+#     def get_queryset(self):
+#         return Follow.objects.filter(user=self.request.user)
+
+
+# class SubscribeAPIView(APIView):
+#     """API вью для подписки и отписки от авторов"""
+#     permission_classes = (AuthorOrReadonly, )
+
+#     def post(self, request, pk=None):
+#         author = get_object_or_404(User, pk=pk)
+#         data = {'user': request.user.id, 'author': author.id}
+#         serializer = FollowSerializer(
+#             data=data, context={'request': request}
+#         )
+#         serializer.is_valid(raise_exception=True)
+#         serializer.save()
+#         return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+#     def delete(self, request, pk):
+#         author = get_object_or_404(User, pk=pk)
+#         follow = get_object_or_404(
+#             Follow, user=request.user, author=author
+#         )
+#         follow.delete()
+#         return Response(status=status.HTTP_204_NO_CONTENT)
